requirements.txt
llama-index-llms-huggingface-api==0.6.0
llama-index-embeddings-huggingface==0.4.0
llama-index-vector-stores-chroma==0.5.2
llama-index-llms-openai==1.102.0
python-dotenv==1.0.0
chromadb>=0.5.17
llama-index-vector-stores-chroma